+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/PartyConstants.cs
@@ -9,5 +9,10 @@
         /// Scene metadata key
         /// </summary>
         public const string METADATA_KEY = "stormancer.party";
+
+        /// <summary>
+        /// Party template id.
+        /// </summary>
+        public const string PARTY_SCENE_TYPE = "party";
     }
 }